fix: cleanup the scripts and helper code

+++ cmd/scripts/k8s/setup-host.sh
@@ -3,20 +3,17 @@
 chroot /host /bin/bash -c '
 #!/bin/bash
 
+# fail with any of the commands fail
+set -e
+
 # Ensure non-interactive mode for package managers
 export DEBIAN_FRONTEND=noninteractive
-
-if [[ $SKIPSETUP -eq 1 ]]; then
-    cd /
-    IS_K8S=1 ./build-start-daemon.sh --systemctl --no-build
-    exit 0
-fi
 
 YUM_PACKAGES=(wget libnet-devel libnl3-devel libcap-devel libseccomp-devel gpgme-devel btrfs-progs-devel buildah criu protobuf protobuf-c protobuf-c-devel protobuf-c-compiler protobuf-compiler protobuf-devel python3-protobuf)
 APT_PACKAGES=(wget libnl-3-dev libnet-dev libbsd-dev libcap-dev pkg-config libgpgme-dev libseccomp-dev libbtrfs-dev buildah libprotobuf-dev libprotobuf-c-dev protobuf-c-compiler protobuf-compiler python3-protobuf)
 
 check_and_install_apt_packages() {
-    apt-get update -y
+    apt-get update
 
     local missing_packages=()
     for pkg in "${APT_PACKAGES[@]}"; do
@@ -25,26 +22,6 @@ check_and_install_apt_packages() {
         fi
     done
 
-    # Check if CRIU is installed
-    if ! dpkg -l | grep -qw "criu"; then
-        case $(uname -m) in
-        x86_64)
-            PACKAGE_URL="https://download.opensuse.org/repositories/devel:/tools:/criu/xUbuntu_22.04/amd64/criu_3.19-4_amd64.deb"
-            OUTPUT_FILE="criu_3.19-4_amd64.deb"
-        aarch64)
-            PACKAGE_URL="https://download.opensuse.org/repositories/devel:/tools:/criu/xUbuntu_22.04/arm64/criu_3.19-4_arm64.deb"
-            OUTPUT_FILE="criu_3.19-4_arm64.deb"
-        *)
-            echo "Unknown platform $(uname -m)"
-            exit 1
-        esac
-
-        wget "$PACKAGE_URL" -O "$OUTPUT_FILE"
-        dpkg -i "$OUTPUT_FILE"
-    else
-        echo "CRIU is already installed"
-    fi
-
     # Install other APT packages if missing
     if [ ${#missing_packages[@]} -gt 0 ]; then
         apt-get install -y -o Dpkg::Options::="--force-confdef" -o Dpkg::Options::="--force-confold" "${missing_packages[@]}"
@@ -52,6 +29,33 @@ check_and_install_apt_packages() {
         echo "All APT packages are already installed"
     fi
 
+    # check if CRIU is installed
+    # note: criu requires some of the deps we install
+    # so only install is after all other packages have been installed
+    if ! dpkg -l | grep -qw "criu"; then
+        case $(uname -m) in
+        x86_64)
+            PACKAGE_URL="https://download.opensuse.org/repositories/devel:/tools:/criu/xUbuntu_22.04/amd64/criu_3.19-4_amd64.deb"
+            OUTPUT_FILE="criu_3.19-4_amd64.deb"
+            ;;
+        aarch64)
+            PACKAGE_URL="https://download.opensuse.org/repositories/devel:/tools:/criu/xUbuntu_22.04/arm64/criu_3.19-4_arm64.deb"
+            OUTPUT_FILE="criu_3.19-4_arm64.deb"
+            ;;
+        *)
+            echo "Unknown platform $(uname -m)"
+            exit 1
+            ;;
+        esac
+
+        if ! test -f "$OUTPUT_FILE"; then
+            # do not make network request if the file is already available
+            wget "$PACKAGE_URL" -O "$OUTPUT_FILE"
+        fi
+        dpkg -i "$OUTPUT_FILE"
+    else
+        echo "$(criu --version) is already installed"
+    fi
 }
 
 check_and_install_yum_packages() {
@@ -66,7 +70,7 @@ check_and_install_yum_packages() {
     if ! rpm -q "criu" &>/dev/null; then
         yum install -y criu
     else
-        echo "CRIU is already installed"
+        echo "$(criu --version) is already installed"
     fi
 
     # Install other YUM packages if missing
@@ -78,45 +82,19 @@ check_and_install_yum_packages() {
     fi
 }
 
-install_criu_ubuntu_2204() {
-    case $(uname -m) in
-        x86 | x86_64)
-            PACKAGE_URL="https://download.opensuse.org/repositories/devel:/tools:/criu/xUbuntu_22.04/amd64/criu_3.19-4_amd64.deb"
-            OUTPUT_FILE="criu_3.19-4_amd64.deb"
-            ;;
-        armv7 | aarch64)
-            PACKAGE_URL="https://download.opensuse.org/repositories/devel:/tools:/criu/xUbuntu_22.04/arm64/criu_3.19-4_arm64.deb"
-            OUTPUT_FILE="criu_3.19-4_arm64.deb"
-            ;;
-        *)
-            echo "Unknown platform $(uname -m)"
-            exit 1
-            ;;
-    esac
-
-    if ! test -f "$OUTPUT_FILE"; then
-        wget "$PACKAGE_URL" -O "$OUTPUT_FILE"
-        dpkg -i "$OUTPUT_FILE"
-    fi
-}
-
 if [ -f /etc/os-release ]; then
     . /etc/os-release
     case "$ID" in
-        debian | ubuntu)
-            check_and_install_apt_packages
-            install_criu_ubuntu_2204
-            ;;
-        rhel | centos | fedora)
-            check_and_install_yum_packages
-            ;;
-        amzn)
-            check_and_install_yum_packages
-            ;;
-        *)
-            echo "Unknown distribution"
-            exit 1
-            ;;
+    debian | ubuntu)
+        check_and_install_apt_packages
+        ;;
+    amzn | rhel | centos | fedora)
+        check_and_install_yum_packages
+        ;;
+    *)
+        echo "Unknown distribution"
+        exit 1
+        ;;
     esac
 elif [ -f /etc/debian_version ]; then
     check_and_install_apt_packages
@@ -127,9 +105,12 @@ else
     exit 1
 fi
 
-systemctl stop cedana.service
+# only try to stop if the service exists
+if systemctl --status-all | grep -Fq "cedana"; then
+    systemctl stop cedana.service
+fi
 
-rm -rf /var/log/cedana*
+rm -rf /var/log/cedana
 
 '
